// src/features/chat/data/enums.ts
export const enum CHAT_ROLES {
  REQUESTER = "REQUESTER",
  ADMIN = "ADMIN",
  TECHNICIAN = "TECHNICIAN",
}

export const enum CHAT_TYPE {
  USER = "USER",
  SYSTEM = "SYSTEM",
}

export const enum CHAT_MSG_DELIVERY_STATUS {
  SENDING = "sending",
  SENT = "sent",
  DELIVERED = "delivered",
  READ = "read",
}